Code change details:
1. Assert error messages modified for better investigation, in case of failure. Now when AssertionError occurs, it will display the exact image name.

2. To handle the following 6 scenarios, six new regions have been implemented:

There are 4 cells displayed in a row.
If the browser is resized, the cells will be arranged as 3 cells in one row.
If the browser is further resized, ultimately the cells will be arranged as 3 cells in 2 rows.
There are 8 cells displayed in 2 rows.
If the browser is resized, the cells will be arranged as 6 cells in one row,
If the browser is further resized, ultimately the cells will be arranged as 6 cells in 3 rows.

3. Added DEFAULT_MOVE_MOUSE_DELAY after a click instance.