Issue 3125: Firefox Home Content - Highlights can be displayed in 1 or 2 rows (#4221)
* Issue 3125: Firefox Home Content - Highlights can be displayed in 1 or 2 rows
Test case scripted and tested for all four operating systems.

* Fixes for intermittent failure in Linux VM. The changes done in this commit are:
1. Changed similarity pattern matching for "home_page_highlights_pattern" image from 60 % to 70%.
2. Replaced youtube by twitter as youtube load time was high due to video content loading, hence test execution time was more.
3. Added FIREFOX_TIMEOUT in "exists".
4. Handled firefox optional message in the footer.
5. Changed resized browser height from 700 to 800.

Tested again in local machines and VMs for all OS. Test Passed.

* Code change details:

1. Assert error messages modified for better investigation, in case of failure. Now when AssertionError occurs, it will display the exact image name.

2. To handle the following 6 scenarios, six new regions have been implemented:

There are 4 cells displayed in a row.
If the browser is resized, the cells will be arranged as 3 cells in one row.
If the browser is further resized, ultimately the cells will be arranged as 3 cells in 2 rows.
There are 8 cells displayed in 2 rows.
If the browser is resized, the cells will be arranged as 6 cells in one row,
If the browser is further resized, ultimately the cells will be arranged as 6 cells in 3 rows.

3. Added DEFAULT_MOVE_MOUSE_DELAY after a click instance.

* Code change details:
The 6 scenarios to create six regions have been implemented using "region_creation_highlights" function. This has reduced code length and optimised code.